Part4 bloglist expansion step 10

+++ controllers/blogs.js
@@ -2,19 +2,21 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
-  
-blogsRouter.post('/', async (request, response, next) => {
+
+blogsRouter.post('/', userExtractor, async (request, response, next) => {
   const body = request.body
 
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token invalid' })
   }
+  const userReq = request.user
   const user = await User.findById(decodedToken.id)
 
   const blog = new Blog({
@@ -25,23 +27,28 @@ blogsRouter.post('/', async (request, response, next) => {
     user: user._id
   })
 
+  console.log(blog.user.toString())
+  console.log(userReq.toString())
   try {
-    const savedBlog = await blog.save()
-    user.blogs = user.blogs.concat(savedBlog._id)
-    await user.save()
-    response.status(201).json(savedBlog)
-  } catch(exception) {
+    if (blog.user.toString() === userReq.toString()) {
+      const savedBlog = await blog.save()
+      user.blogs = user.blogs.concat(savedBlog._id)
+      await user.save()
+      response.status(201).json(savedBlog)
+    } else {
+      response.status(400).json({ error: "incorrect user" })
+    }
+  } catch (exception) {
     response.status(400).end()
   }
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response, next) => {
   try {
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
     const blog = await Blog.findById(request.params.id)
-    console.log(blog.user.toString())
-    console.log(decodedToken.id.toString())
-    if (blog.user.toString() === decodedToken.id.toString()) {
+    const user = request.user
+    if (blog.user.toString() === decodedToken.id.toString() && user.toString() === blog.user.toString()) {
       await Blog.findByIdAndRemove(request.params.id)
       response.status(204).end()
     } else {
